+++ infrastructure/postgres-adapter/src/main/java/onlydust/com/marketplace/api/postgres/adapter/repository/InvoiceViewRepository.java
@@ -7,12 +7,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 import java.util.UUID;
 
 public interface InvoiceViewRepository extends JpaRepository<InvoiceViewEntity, UUID> {
-    List<InvoiceViewEntity> findAllByBillingProfileIdAndStatus(final @NonNull UUID billingProfileId, final @NonNull Invoice.Status status);
-
     Page<InvoiceViewEntity> findAllByBillingProfileIdAndStatusNot(final @NonNull UUID billingProfileId, final @NonNull Invoice.Status exceptStatus,
                                                                   final @NonNull Pageable pageable);
 }
