+++ infrastructure/postgres-adapter/src/main/java/onlydust/com/marketplace/api/postgres/adapter/repository/ProjectRepository.java
@@ -13,9 +13,6 @@ import java.util.UUID;
 
 public interface ProjectRepository extends JpaRepository<ProjectEntity, UUID>, JpaSpecificationExecutor<ProjectEntity> {
 
-    @Query(value = "SELECT slug FROM projects WHERE id = ?1", nativeQuery = true)
-    String getSlugById(UUID id);
-
     Page<ProjectEntity> findAllByNameContainingIgnoreCase(String name, Pageable pageable);
 
     @Query(nativeQuery = true, value = """
