--- project-domain/src/main/java/onlydust/com/marketplace/project/domain/view/CodeReviewState.java
@@ -1,5 +1,0 @@
-package onlydust.com.marketplace.project.domain.view;
-
-public enum CodeReviewState {
-    PENDING, COMMENTED, APPROVED, CHANGES_REQUESTED, DISMISSED;
-}
