--- project-domain/src/main/java/onlydust/com/marketplace/project/domain/view/JuryVoteView.java
@@ -1,4 +1,0 @@
-package onlydust.com.marketplace.project.domain.view;
-
-public record JuryVoteView() {
-}
