--- project-domain/src/main/java/onlydust/com/marketplace/project/domain/view/ShortSponsorView.java
@@ -1,11 +1,0 @@
-package onlydust.com.marketplace.project.domain.view;
-
-import lombok.Builder;
-import lombok.NonNull;
-
-import java.util.UUID;
-
-
-@Builder
-public record ShortSponsorView(@NonNull UUID id, @NonNull String name, String url, @NonNull String logoUrl) {
-}
