--- project-domain/src/main/java/onlydust/com/marketplace/project/domain/view/commitee/VoteView.java
@@ -1,6 +1,0 @@
-package onlydust.com.marketplace.project.domain.view.commitee;
-
-import lombok.NonNull;
-
-public record VoteView(@NonNull String criteria, Integer score) {
-}
